New category for time span filter: last x days Fixed "Error loading library function: GetVolumeInformationByHandleW" if NTFS permissions are copied Fixed command line issues: allow config name without extension, allow multiple directories instead of a config file Reenabled global shortcut F8 to toggle data shown in middle grid Unified error handling on failure to create log directory Do not close batch creation dialog after save Tree view: compress and filter root nodes the same way as regular folder nodes Fixed wrong tool tip being shown if directory name changes Date range selector does not trim year field anymore Show action "do nothing" on mouse-hover for conflicts in middle grid Fixed "Windows Error Code 59: An unexpected network error occurred" New filter pattern: *\* matches all files in sub directories of base directories Fixed "*?" filter sub-sequence Fixed "Cannot convert from the charset 'Unknown encoding (-1)'!" Support Ctrl + A in filter dialog Support large filter lists > 32 kByte Allow to hide file icons Avoid switching monitor when main dialog is maximized on multiple monitor systems Improved huge XML file loading times by a factor of 3000, saving by a factor of 3 Restore grid scroll position after repeated comparisons Show log after sync when non-fatal errors occurred Fixed crash in UTF8 conversion when processing a corrupted ffs_db file Even more pedantic user interface fine-tuning Compiles and runs on openSuse 12.1 Fixed grid page-up/down keys scrolling twice (Linux, wxGTK 2.9.3) Fixed unwanted grid scrolling when toggling middle column (Linux, wxGTK 2.9.3) Fixed middle grid tool tip occasionally going blank (Linux) Support single shift-click to check/set direction of multiple rows Removed gtkmm dependency (Linux) Installer remembers all settings for next installation (local installation only) All executables digitally signed Updated translation files

+++ wx+/button.h
@@ -24,18 +24,18 @@ public:
                  const wxValidator& validator = wxDefaultValidator,
                  const wxString& name         = wxButtonNameStr);
 
-    void setBitmapFront(const wxBitmap& bitmap, unsigned spaceAfter = 0);
+    void setBitmapFront(const wxBitmap& bitmap, size_t spaceAfter = 0);
     void setTextLabel(  const wxString& text);
-    void setBitmapBack( const wxBitmap& bitmap, unsigned spaceBefore = 0);
+    void setBitmapBack( const wxBitmap& bitmap, size_t spaceBefore = 0);
 
 private:
     wxBitmap createBitmapFromText(const wxString& text);
     void refreshButtonLabel();
 
     wxBitmap bitmapFront;
-    unsigned m_spaceAfter;
+    unsigned int m_spaceAfter;
     wxString textLabel;
-    unsigned m_spaceBefore;
+    unsigned int m_spaceBefore;
     wxBitmap bitmapBack;
 };
